pyproject.toml
[tool.poetry]
name = "starnavi"
version = "1.0"
description = "Starnavi test task"
authors = []

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1"
psycopg2-binary = "^2.9"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
